refactor: inline IetfLiveServerTestCase
(there's only one subclass)
